backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cantomeet-notes-backend"
version = "0.1.0"
description = "FastAPI backend for the CantoMeet Notes MVP."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.23.0,<1.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "psycopg[binary]>=3.1.0,<4.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.25.0,<1.0.0",
    "rq>=1.15.0,<2.0.0",
    "redis>=5.0.0,<6.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    "email-validator>=2.0.0,<3.0.0",
    "openai>=1.0.0,<2.0.0",  # For Whisper API support
    "certifi>=2023.0.0",  # SSL certificates for macOS compatibility
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.4.0,<1.0.0",
    "black>=24.0.0",
]

[project.scripts]
cantomeet-backend = "app.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*", "*.egg-info*"]


